+++ lib/dradis-burp.rb
@@ -6,3 +6,5 @@ require 'dradis/plugins/burp'
 
 # Load supporting Burp classes
 require 'burp/issue'
+require 'burp/xml/issue'
+require 'burp/html/issue'